refactor: return inline 변경 및 FileService 이동
