+++ src/main/java/com/bitharmony/comma/album/Image/controller/AlbumImageController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.bitharmony.comma.album.Image.service.AlbumImageService;
-import com.bitharmony.comma.album.file.service.FileService;
+import com.bitharmony.comma.file.service.FileService;
 import com.bitharmony.comma.global.response.GlobalResponse;
 
 import lombok.RequiredArgsConstructor;
